deprecate sorted array
no more extending builtins; doesn't transpile well on older targets
